# contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: driftmail API
  version: 0.1.0
  description: >
    Track 0 Contract. Track A implementiert diese Spec real.
    Track C und F bauen zunächst gegen einen Mock-Server, der exakt
    diese Response-Formate zurückgibt.
servers:
  - url: https://api.driftware.online/v1
security:
  - bearerAuth: []
paths:
...
  /messages:
    get:
      summary: Nachrichten eines Ordners abrufen
      parameters:
        - name: folderId
          in: query
          schema:
            type: string
            format: uuid
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Liste der Nachrichten
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    get:
      summary: Einzelne Nachricht inkl. Security-Analyse und Links
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Nachricht mit Analyse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDetail"
    delete:
      summary: Mail in den Papierkorb verschieben (soft delete)
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: In Papierkorb verschoben
  /messages/{messageId}/permanent:
    delete:
      summary: Mail endgueltig loeschen (nur aus dem Papierkorb heraus moeglich)
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Endgueltig geloescht
  /messages/{messageId}/quarantine:
    post:
      summary: Nachricht manuell in Quarantäne verschieben
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Quarantäne-Eintrag erstellt
  /messages/{messageId}/move:
    post:
      summary: Nachricht in einen anderen Ordner verschieben
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [folderId]
              properties:
                folderId: { type: string, format: uuid }
      responses:
        "200":
          description: Verschoben
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}/unsubscribe:
    post:
      summary: >
        Manuelle Abmeldung ueber RFC-8058 List-Unsubscribe-Header anstoßen.
        Legt einen unsubscribe_actions-Eintrag mit status='pending_confirmation'
        an, nie automatischer Klick auf Links im Mail-Body. 400, wenn die
        Nachricht keinen gueltigen List-Unsubscribe-Header hat. Siehe auch
        WEB_INBOX.md 09.09. "Automatisches Abmelden bei Spam": fuer
        Nachrichten mit classification='spam' (NIE 'phishing') loest das
        Backend denselben Mechanismus zusaetzlich automatisch aus (ohne
        Rueckfrage, status='confirmed' direkt) -- dieser Endpunkt bleibt der
        Weg fuer eine vom User selbst angestoßene Abmeldung bei allen
        anderen Nachrichten mit einem List-Unsubscribe-Header.
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Abmeldung angestoßen
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending_confirmation, confirmed, rejected]
        "400":
          description: Nachricht hat keinen gueltigen List-Unsubscribe-Header
        "404":
          description: Nachricht nicht gefunden
  /messages/{messageId}/summary:
    get:
      summary: KI-Zusammenfassung ("Was wollen die von mir")
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Zusammenfassung
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailSummary"
  /messages/{messageId}/reply-draft:
    post:
      summary: Antwortentwurf generieren lassen
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entwurf
          content:
            application/json:
              schema:
                type: object
                properties:
                  draftText:
                    type: string
  /messages/draft/phishing-check:
    post:
      summary: >
        Prueft einen Mail-Entwurf vor dem Versand auf Phishing-Merkmale
        (harter Block, siehe blocked) sowie nicht-blockierende Warnsignale
        (sensible Daten, Empfaenger-Reputation, riskante Links). Nutzt
        dieselben Signale wie SecurityResult/analyzeMail, angewendet auf den
        eigenen Entwurf statt auf eingehende Mails (siehe WEB_INBOX.md 08.09.).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bodyText: { type: string }
                links:
                  type: array
                  items:
                    type: object
                    properties:
                      displayText: { type: string }
                      actualUrl: { type: string }
                recipientAddress:
                  type: string
                  nullable: true
                  description: >
                    Optional. Ziel-Adresse des Entwurfs, Grundlage für
                    recipientReputation (siehe unten). Ohne dieses Feld bleibt
                    recipientReputation "unknown" (SYNC.md 08.09., kleinere
                    Contract-Ergänzung).
      responses:
        "200":
          description: Ergebnis
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocked:
                    type: boolean
                    description: >
                      true = Phishing-Merkmale erkannt, Versand MUSS verhindert
                      werden (kein Warnen-und-trotzdem-erlauben).
                  reason: { type: string, nullable: true }
                  containsSensitiveData:
                    type: array
                    items: { type: string, enum: [iban, credit_card, other] }
                  recipientReputation:
                    type: string
                    enum: [safe, unknown, flagged]
                  riskyLinks:
                    type: array
                    items:
                      type: object
                      properties:
                        url: { type: string }
                        reason: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Session-Token nach OAuth-Verknuepfung des Mail-Kontos. userId wird
        serverseitig aus dem Token aufgeloest, nie aus Request-Body/Query/Pfad
        vertraut.
  schemas:
...
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        fromAddress: { type: string }
        fromDisplayName: { type: string }
        subject: { type: string }
        receivedAt: { type: string, format: date-time }
        folderId: { type: string, format: uuid }
        classification:
          type: string
          enum: [safe, spam, phishing, unclear]
    MessageDetail:
      allOf:
        - $ref: "#/components/schemas/Message"
        - type: object
          properties:
            bodyText: { type: string }
            security:
              $ref: "#/components/schemas/SecurityResult"
            links:
              type: array
              items:
                $ref: "#/components/schemas/MessageLink"
            quarantine:
              $ref: "#/components/schemas/QuarantineInfo"
            canUnsubscribe:
              type: boolean
              description: >
                true, wenn die Nachricht einen syntaktisch gueltigen
                List-Unsubscribe-Header hat (unabhaengig von classification)
                -- steuert, ob die Client-UI den manuellen Abmelden-Button
                zu POST /messages/{messageId}/unsubscribe anbietet.
    QuarantineInfo:
      type: object
      nullable: true
      description: >
        Nur gesetzt, wenn sich die Nachricht aktuell im Quarantaene-Ordner
        befindet (quarantine-Tabelle in db-schema.sql), sonst null.
        TERMINAL_INBOX.md 08.09. Punkt 6: bisher gab es keinen GET-Weg fuer
        reason/auto_delete_at, obwohl die Spalten seit Track 0 existieren.
      properties:
        reason: { type: string }
        autoDeleteAt: { type: string, format: date-time }
        userReviewed: { type: boolean }
    MessageLink:
      type: object
      properties:
        displayText: { type: string }
        actualUrl: { type: string }
        domainMatchesDisplay: { type: boolean }
        isKnownMalicious: { type: boolean }
    SecurityResult:
      type: object
      properties:
        spfStatus: { type: string, enum: [pass, fail, none] }
        dkimStatus: { type: string, enum: [pass, fail, none] }
        dmarcStatus: { type: string, enum: [pass, fail, none] }
        senderDomainAgeDays: { type: integer, nullable: true }
        domainReputationScore: { type: number, format: float, nullable: true }
        homoglyphDetected: { type: boolean }
        linkMismatchDetected: { type: boolean }
        urgencyLanguageScore: { type: number, format: float, nullable: true }
        containsNewIban: { type: boolean }
        classification: { type: string, enum: [safe, spam, phishing, unclear] }
        spamSubcategory:
          type: string
          nullable: true
          enum: [adult, gambling, generic, marketing]
        ipReputationFlag:
          type: string
          nullable: true
          enum: [clean, known_botnet, unknown]
        heloMismatch: { type: boolean }
        imageToTextRatio: { type: number, format: float, nullable: true }
        confidenceScore: { type: number, format: float }
    MailSummary:
      type: object
      properties:
        summaryText: { type: string }
        actionRequired: { type: boolean }
        actionDescription: { type: string }
        deadline: { type: string, format: date, nullable: true }
        source: { type: string, enum: [on_device, cloud_fallback] }
...